refactor: relocated init_db
